misc: update page Le Me
