--- website/src/pages/this/styles.module.css
@@ -1,26 +1,0 @@
-.thisHead {
-    background-color: var(--ifm-site-dark) !important;
-    flex-direction: column;
-    /* padding: 3.6rem 0 1rem; */
-    padding: 3.6rem 0 1rem;
-}
-
-.thisTitle {
-    font-size: 50px;
-    font-weight: 600;
-    margin-bottom: 0.1rem;
-}
-
-
-.thisBody {
-    /* background-color: var(--ifm-color-primary) !important; */
-    background-color: var(--ifm-site-dark) !important;
-    flex-direction: column;
-    font-size: 20px;
-    font-weight: 300;
-}
-
-.thisAvatar {
-    max-height: 196px;
-    max-width: 196px;
-}
